update api cinema, api room

+++ movie-theater-api/database/migrations/2024_11_19_022837_create_movies_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->bigIncrements('movie_id');
             $table->string('movie_name');
             $table->text('summary')->nullable();
-            $table->tinyInteger('duration');
+            $table->integer('duration');
             $table->date('release_date');
             $table->string('author')->nullable();
             $table->string('actor')->nullable();
