+++ movie-theater-api/database/migrations/2024_11_19_033552_create_foods_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->bigIncrements('food_id');
             $table->string('food_name', 255);
             $table->float('price');
-            $table->tinyInteger('stock_quantity');
+            $table->integer('stock_quantity');
             $table->string('image_food')->nullable();
             $table->text('description')->nullable();
             $table->boolean('is_enabled')->default(true);
